refactor(file structure): merge modules/mouse.mjs to features/reading-list/filter.js
